adding rat back animations

+++ Assets/Scripts/Rat/RatController3D.cs
@@ -136,7 +136,7 @@ public class RatController3D : MonoBehaviour
             groundForward = forwardVector;
             // set animator controller for sprite look direction
             int direction = WorldSprite.getSpriteLookDirectionTest(groundForward);
-            animator.SetInteger("direction", direction);
+            animator.SetFloat("direction", direction);
         }
 
         moveVector.Normalize();
